frontend/requirements.txt
streamlit
requests
pandas
Pillow
boto3
python-dotenv
